+++ changelog.md
@@ -98,3 +98,4 @@
 * Support for ENUM datatype
 * user.name field not unique
 * Improved debug info printed to console.log on errors in startup and routes
+* Record reference to overlook in all routes (so can be accessed by route init functions)