+++ changelog.md
@@ -100,3 +100,4 @@
 * Improved debug info printed to console.log on errors in startup and routes
 * Record reference to overlook in all routes (so can be accessed by route init functions)
 * Account namespace including function for user to update details and change password
+* Markdown compilation support